initial try at moving sighting logic to model.. broken
